Refactor: Post 생성 시 ChatRoom 추가 수정중

+++ src/main/java/com/example/gamemate/domain/post/dto/PostUpdateDTO.java
@@ -12,7 +12,7 @@ public class PostUpdateDTO {
     private String status;
 
     @NotNull
-    private Integer mateCnt;
+    private Long mateCnt;
 
     @NotNull
     private String mateContent;
@@ -25,5 +25,4 @@ public class PostUpdateDTO {
     private BigDecimal latitude;
 
     private BigDecimal longitude;
-
 }
